requirements.txt
mcpi
numpy
matplotlib
pillow
opencv-python